Added and exported method `test`
This commit renames the primary function to `parse`, adds a method named `test`
and assigns `module.exports` to `parse.

The `test` method lets one test a string for the existence of front-matter. It
takes a string, and returns true if front-matter is found and false if front-
matter is not found.

The primary front-matter function is renamed to parse, but does not change the
way the module is used.

The `test` method will differentiate between strings with empty front-matter and
strings that don't have any front-matter which differentiates it from the
primary `parse` function which returns an attributes object no matter the
presence of a front-matter header.